fix empty previous folder

+++ linux/config/set-repositories-path.sh
@@ -10,10 +10,12 @@ mkdir $repositories
 chown -R $config_user:$config_group $repositories
 chmod -R 777 $repositories
 
-cp repositories-hooks/* $repositories
-cp $config_repositories/* $repositories
+if [ -d $config_repositories ]; then
+	cp -r repositories-hooks/* $repositories
+	cp -r $config_repositories/* $repositories
 
-rm -rf $config_repositories
+	rm -rf $config_repositories
+fi
 
 gman="./../gman.sh"
 writeline="./../utils/wline.sh"
